M182 break up huge and growing test file
